requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
yt-dlp==2024.12.13
websockets==13.1